requirements.txt
Flask==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
gunicorn==20.1.0